# swagger.yaml
openapi: 3.0.3
info:
  title: Anony API
  version: 1.0.0
  description: A simple social media api, created with expressjs, sql and sequelize ORM
  contact:
    name: GitHub Repository
    url: https://github.com/naveen-2111-dev/seq-soc
servers:
  - url: http://localhost:8080
...
security:
  - accessTokenCookie: []
paths:
...
  /api/posts/create:
    post:
      tags:
        - Posts
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreateRequest'
      responses:
        '200':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/posts/update/{postId}:
    put:
      tags:
        - Posts
      summary: Update a post  
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUpdateRequest'
      responses:
        '200':
          description: Post updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/posts/visibility/{postId}:
    patch:
      tags:
        - Posts
      summary: Change post visibility
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisibilityUpdateRequest'
      responses:
        '200':
          description: Visibility updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/posts/all:
    get:
      tags:
        - Posts
      summary: Get all visible posts
      responses:
        '200':
          description: Posts fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostArrayResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/posts/trending:
    get:
      tags:
        - Posts
      summary: Get trending posts
      responses:
        '200':
          description: Trending posts fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostWithAuthorArrayResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/posts/visibility/{visibility}:
    get:
      tags:
        - Posts
      summary: Get posts by visibility
      parameters:
        - name: visibility
          in: path
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: Posts fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostArrayResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/posts:
    get:
      tags:
        - Posts
      summary: Get posts for the authenticated user
      responses:
        '200':
          description: Posts fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostArrayResponse'
        '400':
          description: User ID is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/posts/{postId}:
    get:
      tags:
        - Posts
      summary: Get a post by id
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Post fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/comments/post/{postId}:
    get:
      tags:
        - Comments
      summary: Get comments for a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Comments fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentArrayResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/follow/follow:
    post:
      tags:
        - Follow
      summary: Follow a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FollowRequest'
      responses:
        '200':
          description: Follow created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/follow/unfollow:
    delete:
      tags:
        - Follow
      summary: Unfollow a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FollowRequest'
      responses:
        '200':
          description: Follow removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnfollowResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/follow/followers:
    get:
      tags:
        - Follow
      summary: Get followers for the authenticated user
      responses:
        '200':
          description: Followers fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowArrayResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/follow/following:
    get:
      tags:
        - Follow
      summary: Get following for the authenticated user
      responses:
        '200':
          description: Following fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FollowArrayResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/like/like:
    post:
      tags:
        - Like
      summary: Like a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LikeRequest'
      responses:
        '200':
          description: Post liked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '400':
          description: User ID and Post ID are required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/like/unlike:
    delete:
      tags:
        - Like
      summary: Unlike a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LikeRequest'
      responses:
        '200':
          description: Post unliked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '400':
          description: User ID and Post ID are required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/save/save:
    post:
      tags:
        - Save
      summary: Save a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveRequest'
      responses:
        '200':
          description: Post saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveResponse'
        '400':
          description: User ID is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/save/unsave:
    delete:
      tags:
        - Save
      summary: Unsave a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveRequest'
      responses:
        '200':
          description: Post unsaved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BaseResponse'
        '400':
          description: User ID is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Saved post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/save/user:
    get:
      tags:
        - Save
      summary: Get saved posts for the authenticated user
      responses:
        '200':
          description: Saved posts fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedPostArrayResponse'
        '400':
          description: User ID is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    accessTokenCookie:
      type: apiKey
      in: cookie
      name: accessToken
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refreshToken
  schemas:
...
    PostCreateRequest:
      type: object
      required:
        - content
        - userId
      properties:
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        userId:
          type: integer
    PostUpdateRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
    VisibilityUpdateRequest:
      type: object
      required:
        - visibility
      properties:
        visibility:
          type: boolean
...
    FollowRequest:
      type: object
      required:
        - followingId
      properties:
        followingId:
          type: integer
    LikeRequest:
      type: object
      required:
        - postId
      properties:
        postId:
          type: integer
    SaveRequest:
      type: object
      required:
        - postId
      properties:
        postId:
          type: integer
...
    UserPublic:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        fullName:
          type: string
        profileImage:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Follow:
      type: object
      properties:
        id:
          type: integer
        followerId:
          type: integer
        followingId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SavePost:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
          readonly: true
        postId:
          type: integer
        User:
          $ref: '#/components/schemas/UserPublic'
        Post:
          $ref: '#/components/schemas/Post'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Post:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        visibility:
          type: boolean
        upVotes:
          type: integer
        downVotes:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PostWithAuthor:
      allOf:
        - $ref: '#/components/schemas/Post'
        - type: object
          properties:
            author:
              $ref: '#/components/schemas/UserPublic'
    Comment:
      type: object
      properties:
        id:
          type: integer
        postId:
          type: integer
        userId:
          type: integer
          nullable: true
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    BaseResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            error:
              type: object
              additionalProperties: true
...
    PostResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Post'
    PostArrayResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Post'
    PostWithAuthorArrayResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/PostWithAuthor'
...
    CommentArrayResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Comment'
...
    FollowResponse:
      type: object
      properties:
        id:
          type: integer
        followerId:
          type: integer
        followingId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FollowArrayResponse:
      type: array
      items:
        $ref: '#/components/schemas/Follow'
    UnfollowResponse:
      type: integer
      description: Number of rows deleted
    SaveResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/SavePost'
    SavedPostArrayResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/SavePost'
...